Add support for Full Site Editing.
Full Site Editing is now supported in WordPress >= 5.8 with this theme.

The template editing mode is a way to edit the website without the complexity of the site editor interface.
It is more limited than the site editor because you can not select or navigate between templates in this view.

You access the template editing mode via the block editor.

Create a new post or page. Next, open the document settings sidebar and locate the Template panel below Status & visibility.

Here you will find information about the current template and a list of existing templates to choose from.

Create a new template by selecting the New link.

Edit and save the template in the same way as in the site editor.